Silence some warnings caused by [26825]
git-svn-id: http://svn.mythtv.org/svn/trunk@26826 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythui/mythuibuttontree.cpp
@@ -415,7 +415,7 @@ void MythUIButtonTree::SwitchList(bool right)
     if (right)
     {
         if ((m_activeListID < m_visibleLists - 1) &&
-            (m_activeListID < m_buttonlists.count() - 1))
+            (m_activeListID < (uint)m_buttonlists.count() - 1))
             m_activeListID++;
         else if (m_currentNode && m_currentNode->visibleChildCount() > 0)
         {
@@ -444,7 +444,7 @@ void MythUIButtonTree::SwitchList(bool right)
     {
         if (m_activeList)
             m_activeList->Deselect();
-        if (m_activeListID < m_buttonlists.count())
+        if (m_activeListID < (uint)m_buttonlists.count())
         {
             m_activeList = m_buttonlists[m_activeListID];
             m_activeList->Select();
